massive refactor, split everything up
client was fully refactored, everything now has a
relevant data type and was separated into smaller functions.

server was partially refactored, it was just split into
more files. it should eventually be organized like client
