+++ .gitignore
@@ -1,3 +1,4 @@
+.idea/
 testing/
 client/client
 server/server
