+++ client/network/cmd.go
@@ -1,16 +1,19 @@
-package cmd
+package network
 
 import (
 	"encoding/base64"
+	"errors"
+	"github.com/malcolmseyd/natpunch-go/client/util"
 	"log"
 	"os/exec"
 	"strconv"
 	"strings"
-
-	"github.com/malcolmseyd/natpunch-go/client/network"
 )
 
-const persistentKeepalive = "25"
+var (
+	// ErrIfaceDown is returned when the Wireguard interface is down
+	ErrIfaceDown = errors.New("client/net: Wireguard interface down")
+)
 
 // RunCmd runs a command and returns the output, returning any errors
 func RunCmd(command string, args ...string) (string, error) {
@@ -21,13 +24,31 @@ func RunCmd(command string, args ...string) (string, error) {
 	return string(outBytes), nil
 }
 
+func SetIfaceUp(ifaceName string) (err error) {
+	_, err = RunCmd("wg-quick", "up", ifaceName)
+	return
+}
+
+// NewClient creates a new network.Client object for the Wireguard interface specified
+func NewClient(ifaceName string, server *Server) (client Client, err error) {
+	// get the source ip that we'll send the packet from
+	client.IP = GetClientIP(server.Addr.IP)
+
+	// get info about the Wireguard config
+	client.Port = GetClientPort(ifaceName)
+	client.Pubkey = GetClientPubkey(ifaceName)
+	client.Privkey = GetClientPrivkey(ifaceName)
+
+	return
+}
+
 // GetClientPort gets the client's listening port for Wireguard
 func GetClientPort(iface string) uint16 {
 	output, err := RunCmd("wg", "show", iface, "listen-port")
 	if err != nil {
 		log.Fatalln("Error getting listen port:", err)
 	}
-	// guaranteed castable to uint16, as ports are only 2 bytes and positive
+	// guaranteed to cast into uint16, as ports are only 2 bytes and positive
 	port, err := strconv.ParseUint(strings.TrimSpace(output), 10, 16)
 	if err != nil {
 		log.Fatalln("Error parsing listen port:", err)
@@ -36,17 +57,17 @@ func GetClientPort(iface string) uint16 {
 }
 
 // GetPeers returns a list of peers on the Wireguard interface
-func GetPeers(iface string) []string {
+func GetPeers(iface string) ([]Peer, error) {
 	output, err := RunCmd("wg", "show", iface, "peers")
 	if err != nil {
-		log.Fatalln("Error getting peers:", err)
+		return nil, err
 	}
-	return strings.Split(strings.TrimSpace(output), "\n")
+	peerKeysStr := strings.Split(strings.TrimSpace(output), "\n")
+	return util.MakePeerSlice(peerKeysStr)
 }
 
-// GetClientPubkey returns the publib key on the Wireguard interface
-func GetClientPubkey(iface string) network.Key {
-	var keyArr [32]byte
+// GetClientPubkey returns the public key on the Wireguard interface
+func GetClientPubkey(iface string) (key Key) {
 	output, err := RunCmd("wg", "show", iface, "public-key")
 	if err != nil {
 		log.Fatalln("Error getting client pubkey:", err)
@@ -55,32 +76,32 @@ func GetClientPubkey(iface string) network.Key {
 	if err != nil {
 		log.Fatalln("Error parsing client pubkey:", err)
 	}
-	copy(keyArr[:], keyBytes)
-	return network.Key(keyArr)
+	copy(key[:], keyBytes)
+	return
 }
 
-// GetClientPrivkey returns the private key on the Wireguard interface
-func GetClientPrivkey(iface string) network.Key {
-	var keyArr [32]byte
+// GetClientPrivkey returns the public key on the Wireguard interface
+func GetClientPrivkey(iface string) (key Key) {
 	output, err := RunCmd("wg", "show", iface, "private-key")
 	if err != nil {
-		log.Fatalln("Error getting client privkey:", err)
+		log.Fatalln("Error getting client pubkey:", err)
 	}
 	keyBytes, err := base64.StdEncoding.DecodeString(strings.TrimSpace(output))
 	if err != nil {
 		log.Fatalln("Error parsing client privkey:", err)
 	}
-	copy(keyArr[:], keyBytes)
-	return network.Key(keyArr)
+	copy(key[:], keyBytes)
+	return
 }
 
 // SetPeer updates a peer's endpoint and keepalive with `wg`. keepalive is in seconds
-func SetPeer(peer *network.Peer, keepalive int, iface string) {
+func UpdatePeer(peer *Peer, keepalive int, iface string) (err error) {
 	keyString := base64.StdEncoding.EncodeToString(peer.Pubkey[:])
-	RunCmd("wg",
+	_, err = RunCmd("wg",
 		"set", iface,
 		"peer", keyString,
 		"persistent-keepalive", strconv.Itoa(keepalive),
 		"endpoint", peer.IP.String()+":"+strconv.FormatUint(uint64(peer.Port), 10),
 	)
+	return
 }
