+++ go.mod
@@ -5,8 +5,10 @@ go 1.15
 require (
 	github.com/flynn/noise v0.0.0-20180327030543-2492fe189ae6
 	github.com/google/gopacket v1.1.18
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/ogier/pflag v0.0.1
 	github.com/vishvananda/netlink v1.1.0
 	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a
 	golang.org/x/net v0.0.0-20200822124328-c89045814202
+	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
 )
